requirements.txt
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
PyPDF2==3.0.1
python-docx==0.8.11
langchain==0.1.0
langchain-community==0.0.13
pydantic>=2.0.0
sentence-transformers>=2.2.2
python-dotenv==1.0.0
httpx>=0.24.0
torch>=2.0.0
chromadb>=0.4.22